+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for compiling shaders in unsaved text editor
+- Added animation on shader compilation
 
 ### Changed
 - Reusing shader output window to not spawn many windows